restructure files out of main.rs
