+++ Cargo.toml
@@ -3,8 +3,7 @@
 members = [
     "blackheap",
     "blackheap-benchmarker",
-    "blackheap-core",
-    "blackheap-analyzer"
+    "blackheap-core"
 ]
 
 resolver = "2"
