+++ blackheap/Cargo.toml
@@ -10,7 +10,6 @@ authors = ["Lars Quentin <lars@lquenti.de>"]
 readme = "./README.md"
 
 [dependencies]
-blackheap-analyzer = { path = "../blackheap-analyzer" }
 blackheap-benchmarker = { path = "../blackheap-benchmarker" }
 blackheap-core = { path = "../blackheap-core" }
 clap = { version = "4.5", features = ["derive"] }
